wip: split files of bst
