+++ src/rust/bst/src/query/common.rs
@@ -1,0 +1,29 @@
+/// Comparison operators for queries.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Operator {
+    Eq,
+    Ne,
+    Gt,
+    Lt,
+    Ge,
+    Le,
+}
+
+/// A condition in a query, consisting of a field name, an operator, and a key value.
+///
+/// The key value is stored as a byte vector, obtained via ByteSerializable::to_bytes.
+#[derive(Debug, Clone)]
+pub struct QueryCondition {
+    /// The field identifier (e.g., "id", "name", etc.)
+    pub field: String,
+    /// The comparison operator.
+    pub operator: Operator,
+    /// The key value as a byte vector (obtained via ByteSerializable::to_bytes).
+    pub key: Vec<u8>,
+}
+
+/// A query consisting of one or more conditions.
+#[derive(Debug, Clone)]
+pub struct Query {
+    pub conditions: Vec<QueryCondition>,
+}
