+++ src/rust/bst/src/query/fs.rs
@@ -1,0 +1,99 @@
+use crate::sorted_index::{SearchableIndex, ValueOffset};
+use std::collections::HashMap;
+use std::collections::HashSet;
+
+use super::common::Query;
+use super::Operator;
+/// A multi-index that maps field names to their corresponding indices.
+pub struct MultiIndex {
+    /// A mapping from field names to their corresponding index.
+    pub indices: HashMap<String, Box<dyn SearchableIndex>>,
+}
+
+impl Default for MultiIndex {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl MultiIndex {
+    /// Create a new, empty multi-index.
+    pub fn new() -> Self {
+        Self {
+            indices: HashMap::new(),
+        }
+    }
+
+    /// Add an index for a field.
+    pub fn add_index(&mut self, field_name: String, index: Box<dyn SearchableIndex>) {
+        self.indices.insert(field_name, index);
+    }
+
+    /// Execute a query against the multi-index.
+    ///
+    /// Returns a vector of offsets for records that match all conditions in the query.
+    pub fn query(&self, query: Query) -> Vec<ValueOffset> {
+        let mut candidate_sets: Vec<HashSet<ValueOffset>> = Vec::new();
+
+        for condition in query.conditions {
+            if let Some(index) = self.indices.get(&condition.field) {
+                let offsets: Vec<ValueOffset> = match condition.operator {
+                    Operator::Eq => {
+                        // Exactly equal.
+                        index.query_exact_bytes(&condition.key)
+                    }
+                    Operator::Gt => {
+                        // Keys strictly greater than the boundary:
+                        // Use query_range_bytes(Some(key), None) and remove those equal to key.
+                        let offsets = index.query_range_bytes(Some(&condition.key), None);
+                        let eq = index.query_exact_bytes(&condition.key);
+                        offsets.into_iter().filter(|o| !eq.contains(o)).collect()
+                    }
+                    Operator::Ge => {
+                        // Keys greater than or equal.
+                        index.query_range_bytes(Some(&condition.key), None)
+                    }
+                    Operator::Lt => {
+                        // Keys strictly less than the boundary.
+                        index.query_range_bytes(None, Some(&condition.key))
+                    }
+                    Operator::Le => {
+                        // Keys less than or equal to the boundary:
+                        // Union the keys that are strictly less and those equal to the boundary.
+                        let mut offsets = index.query_range_bytes(None, Some(&condition.key));
+                        let eq = index.query_exact_bytes(&condition.key);
+                        offsets.extend(eq);
+                        // Remove duplicates by collecting into a set.
+                        let set: HashSet<ValueOffset> = offsets.into_iter().collect();
+                        set.into_iter().collect()
+                    }
+                    Operator::Ne => {
+                        // All offsets minus those equal to the boundary.
+                        let all: HashSet<ValueOffset> =
+                            index.query_range_bytes(None, None).into_iter().collect();
+                        let eq: HashSet<ValueOffset> = index
+                            .query_exact_bytes(&condition.key)
+                            .into_iter()
+                            .collect();
+                        all.difference(&eq).cloned().collect::<Vec<_>>()
+                    }
+                };
+                candidate_sets.push(offsets.into_iter().collect());
+            }
+        }
+
+        if candidate_sets.is_empty() {
+            return vec![];
+        }
+
+        // Intersect candidate sets.
+        let mut intersection: HashSet<ValueOffset> = candidate_sets.first().unwrap().clone();
+        for set in candidate_sets.iter().skip(1) {
+            intersection = intersection.intersection(set).cloned().collect();
+        }
+
+        let mut result: Vec<ValueOffset> = intersection.into_iter().collect();
+        result.sort();
+        result
+    }
+}
