+++ src/rust/bst/src/query/mod.rs
@@ -1,0 +1,10 @@
+mod common;
+mod fs;
+#[cfg(feature = "http")]
+mod http;
+mod stream;
+pub use common::*;
+pub use fs::*;
+#[cfg(feature = "http")]
+pub use http::*;
+pub use stream::*;
